Move MemcachedClientMethods to its own class
